Moved some checks into NeighbourService
